rename AppTypes to Constants, finally
