+++ src/components/MenuBar.tsx
@@ -4,7 +4,6 @@ import { Screen, AppState } from '../Types'
 
 
 import '../styles/MenuBar.css'
-// import { reportEvent } from '../misc'; // later
 
 interface MenuBarProperties {
   state : AppState
@@ -84,98 +83,3 @@ export default function MenuBar (props : MenuBarProperties) : JSX.Element {
     </div>
   )
 }
-
-// function dehydrateBox (box : BoxState) : BoxState {
-//   const { type } : BoxState = box
-
-//   if (type === BoxType.EXPRESSION) {
-
-//     return {
-//       ...box,
-//       ast : null as any, // TODO: don't
-//       history : [], // TODO: don't
-//       isRunning : false,
-//       breakpoints : [], // TODO: solve how to don't
-//       timeoutID : undefined,
-//     }
-//   }
-
-//   return box
-// }
-
-// function dehydrate (state : AppState) : AppState {
-//   return {
-//     ...state,
-//     submittedBoxes : state.submittedBoxes.map(dehydrateBox)
-//   }
-// }
-
-// function hydrateBox (box : BoxState, macroTable : MacroMap) : BoxState {
-//   const { type } : BoxState = box
-  
-//   if (type === BoxType.EXPRESSION) {
-//     const { singleLetterNames } = box as UntypedLambdaState
-//     const ast : AST = parseExpression((box as UntypedLambdaState).expression, { macroTable, singleLetterNames })
-
-//     return {
-//       ...box,
-//       ast,
-//       history : [ { ast, lastReduction : None, step : 0, message : '', isNormalForm : false } ],
-//     }
-//   }
-
-//   return box
-// }
-
-// function hydrate (dehydrated : AppState) : AppState {
-//   const { macroTable } = dehydrated
-//   const config = { macroTable }
-
-//   return {
-//     ...dehydrated,
-//     submittedBoxes : dehydrated.submittedBoxes.map((box) => hydrateBox(box, macroTable))
-//   }
-// }
-
-// interface Config {
-//   singleLetterNames : boolean
-//   macroTable : MacroMap
-// }
-
-// function parseExpression (expression : string, config : Config) : AST {
-//   const { singleLetterNames : singleLetterVars, macroTable } : Config = config
-  
-//   const tokens : Array<Token> = tokenize(expression, { lambdaLetters : ['λ'], singleLetterVars })
-//   const ast : AST = parse(tokens, macroTable)
-
-//   return ast
-// }
-
-
-// function onFiles (event : ChangeEvent<HTMLInputElement>, onImport : (notebook : NotebookState) => void) : void {
-//   const { target : { files } } = event
-//   if (files === null) {
-//     return
-//   }
-
-//   const file : File = files[0]
-//   const reader : FileReader = new FileReader
-//   reader.onload = (event : Event) => {
-//     const notebook : NotebookState = JSON.parse(reader.result as string)
-
-//     onImport(decodeNotebook(notebook))
-
-//     // onImport(hydrate(state))
-//     // reportEvent('Import notebook', `Notebook named ${ file.name }`, '')
-//   }
-
-//   reader.readAsText(file) 
-// }
-
-// function createURL (content : string) : string {
-//   const data = new Blob([ content ], {
-//     type: 'application/json'
-//   })
-
-//   return window.URL.createObjectURL(data);
-// }
