+++ src/screens/Notebook.tsx
@@ -1,8 +1,3 @@
-// This component needs to be able to specify which Boxes are allowed inside
-// It will also have other settings - maybe something like lock - if it's locked, you can not modify it
-// Exam mode will feature the build of the Frontend which will not import any of the Evaluation Boxes
-
-
 import React, { PureComponent } from 'react'
 import CreateBox from '../components/CreateBox'
 import { BoxType, NotebookState, BoxState } from '../Types'
@@ -26,14 +21,8 @@ export default class Notebook extends PureComponent<Props> {
     this.updateBoxState = this.updateBoxState.bind(this)
     this.makeActive = this.makeActive.bind(this)
     this.onBlur = this.onBlur.bind(this)
-    this.createBoxFromURL = this.createBoxFromURL.bind(this)
-    this.setBoxState = this.setBoxState.bind(this)
-    this.updateURL = this.updateURL.bind(this)
   }
 
-  // componentDidMount () : void {
-  //   this.createBoxFromURL()
-  // }
 
   render () {
     const { state } = this.props
@@ -76,97 +65,6 @@ export default class Notebook extends PureComponent<Props> {
     )
   }
 
-  /**
-   * URL params:
-   *  type : untyped-lambda | markdown | lisp | ... 
-   *  specific : {
-   *    SLI : nul
-   *    strategy : normal | applicative | simplyfied
-   *  }
-   *  source : string
-   */
-  createBoxFromURL () {
-    const urlSearchParams : URLSearchParams = new URL(window.location.toString()).searchParams
-    const type : string | null = urlSearchParams.get('type')
-
-    if (type === null) {
-      return
-    }
-
-    // switch (type) {
-    //   case BoxType.UNTYPED_LAMBDA: {
-
-    //   }
-        
-    //   default:
-    //     break;
-    // }
-    
-    // console.log(window.location.toString())
-    // console.log(urlSearchParams.get('type'))
-    // console.log(urlSearchParams.get('source'))
-
-
-
-
-    // const hash : string = decodeURI(window.location.hash.substring(1))
-    // const isExercise : boolean = hash.indexOf('exercise:') !== -1
-
-    // const expression : string = isExercise ? hash.substring(9) : hash
-
-    // if (expression === '') {
-    //   // return
-    // }
-
-    // const box : BoxState = {
-    //   type : BoxType.EXPRESSION,
-    //   __key : Date.now().toString(),
-    //   expression : '',
-    //   ast : null,
-    //   history : [],
-    //   isRunning : false,
-    //   breakpoints : [],
-    //   timeoutID : undefined,
-    //   timeout : 10,
-    //   isExercise : isExercise,
-    //   strategy : this.getActiveStrategy(),
-    //   singleLetterNames : this.getActiveSingleLetterNames(),
-    //   standalones : this.getActiveStandalones(),
-    //   editor : {
-    //     placeholder : PromptPlaceholder.INIT,
-    //     content : expression,
-    //     caretPosition : expression.length,
-    //     syntaxError : null,
-    //   }
-    // }
-
-    // this.setState({
-    //   ...this.state,
-    //   submittedBoxes : [ box ],
-    //   activeBoxIndex : 0,
-    // })
-  }
-
-  setBoxState (index : number, boxState : BoxState) : void {
-    // const { submittedBoxes } = this.state
-    
-    // const expression : string = boxState.type === BoxType.EXPRESSION ? boxState.editor.content || (boxState as EvaluationState).expression : '' // TODO: DIRTY DIRTY BIG TIME
-    // const expPrefix : string = boxState.type === BoxType.EXPRESSION && (boxState as EvaluationState).isExercise ? 'exercise:' : '' 
-    
-    // history.pushState({}, "page title?", "#" + expPrefix + encodeURI(expression))
-
-    // // TODO: doresit update URL // ted uz to docela dobry je
-
-    // // TODO: consider immutability
-    // submittedBoxes[index] = boxState
-
-    // this.setState({
-    //   ...this.state,
-    //   submittedBoxes,
-    // })
-
-  }
-
   insertBefore (index : number, box : BoxState) : void {
     const { boxList } = this.props.state
 
@@ -206,8 +104,6 @@ export default class Notebook extends PureComponent<Props> {
     // console.log('UPDATING BOX STATE')
     const { boxList } = this.props.state
     boxList[index] = { ...box }
-
-    this.updateURL(box)
 
 
     this.props.updateNotebook({ boxList : [...boxList], activeBoxIndex : index })
@@ -256,8 +152,6 @@ export default class Notebook extends PureComponent<Props> {
           break
       }
 
-      this.updateURL(boxList[index])
-
       this.props.updateNotebook({ activeBoxIndex : index, focusedBoxIndex : index, boxList })
     }
   }
@@ -293,34 +187,5 @@ export default class Notebook extends PureComponent<Props> {
     }
 
     this.props.updateNotebook({ boxList, focusedBoxIndex : undefined })
-  }
-
-  updateURL (box : BoxState) : void {
-    return
-    // switch (box.type) {
-    //   case BoxType.MARKDOWN : {
-    //     const searchParams : URLSearchParams = new URL(window.document.location.toString()).searchParams
-
-    //     searchParams.set('type', BoxType.MARKDOWN)
-    //     searchParams.set('source', encodeURI((box as NoteState).editor.content))
-
-    //     window.history.pushState(null, '', '?' + searchParams.toString())
-    //     break;
-    //   }
-
-    //   case BoxType.UNTYPED_LAMBDA : {
-    //     const searchParams : URLSearchParams = new URL(window.document.location.toString()).searchParams
-
-    //     searchParams.set('type', BoxType.UNTYPED_LAMBDA)
-    //     searchParams.set('source', encodeURI((box as UntypedLambdaState).editor.content))
-
-    //     window.history.pushState(null, '', '?' + searchParams.toString())
-
-    //     break;
-    //   }
-    
-    //   default:
-    //     break;
-    // }
   }
 }
